Adding spearmint handler.

+++ xtrack2_spearmint.py
@@ -1,6 +1,7 @@
 import numpy as np
+import spearmint
 
 
 def main(job_id, params):
     print params
-    return 5 + 10 * np.random.randn(1)
+    return float(5 + 10 * np.random.randn(1))
